# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APIs For Thrift Store",
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/categories": {
            "get": {
                "tags": [
                    "Category"
                ],
                "summary": "Lấy danh sách tất cả các danh mục",
                "operationId": "d9bf255011b42d56bf306832dc3366c8",
                "responses": {
                    "200": {
                        "description": "Danh sách danh mục đã được lấy thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 3
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Nhựa HDPE"
                                            },
                                            "type": {
                                                "type": "string",
                                                "example": "material"
                                            },
                                            "parent_id": {
                                                "type": "integer",
                                                "example": 1
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Không có danh mục nào",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Hiện tại không có danh mục nào."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Lỗi khi lấy danh sách danh mục",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Đã xảy ra lỗi khi lấy danh sách danh mục"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Chi tiết lỗi"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 500
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/category/add": {
            "post": {
                "tags": [
                    "Category"
                ],
                "summary": "Tạo mới một danh mục",
                "operationId": "dc07f16d401a4d7794e899a9e51045de",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Nhựa PP"
                                    },
                                    "type": {
                                        "type": "string",
                                        "example": "material"
                                    },
                                    "parent_id": {
                                        "type": "integer",
                                        "example": 1,
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tạo danh mục thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Tạo danh mục thành công"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 7
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Nhựa PP"
                                                },
                                                "type": {
                                                    "type": "string",
                                                    "example": "material"
                                                },
                                                "parent_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 201
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Tạo danh mục thất bại",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Tạo danh mục thất bại"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Chi tiết lỗi"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 500
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/category/{id}": {
            "get": {
                "tags": [
                    "Category"
                ],
                "summary": "Lấy thông tin chi tiết của một danh mục",
                "operationId": "eabd5d467c0b1c671acaedf17181501f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID của danh mục cần lấy thông tin",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 3
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lấy danh mục thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 3
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Nhựa HDPE"
                                        },
                                        "type": {
                                            "type": "string",
                                            "example": "material"
                                        },
                                        "parent_id": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Không tìm thấy danh mục",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Lấy danh mục thất bại"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Không tìm thấy danh mục với ID: 3"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 500
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Lỗi khi lấy thông tin danh mục",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Lấy danh mục thất bại"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Chi tiết lỗi"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 500
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/category/delete/{id}": {
            "delete": {
                "tags": [
                    "Category"
                ],
                "summary": "Xóa một danh mục",
                "operationId": "c8d38b92325370b0b797ce15752d05c2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID của danh mục cần xóa",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 3
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Xóa danh mục thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Xóa danh mục thành công"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Không tìm thấy danh mục",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Xóa danh mục thất bại"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Không tìm thấy danh mục với ID: 3"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 500
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Xóa danh mục thất bại",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Xóa danh mục thất bại"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Chi tiết lỗi"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 500
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}